Refactor offset table writing
